requirements-onnx.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
flatbuffers==2.0
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.1
install==1.3.5
ninja==1.10.2.3
numpy==1.22.2
onnx==1.11.0
onnxruntime==1.10.0
Pillow==9.0.1
protobuf==3.19.4
psutil==5.9.0
pyspng==0.1.0
requests==2.27.1
torch==1.10.2+cpu
torchaudio==0.10.2+rocm4.1
torchvision==0.11.3+rocm4.2
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
